% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_wiener_maintenance_paths}
\alias{sim_wiener_maintenance_paths}
\title{Simulate Multiple Wiener Degradation Paths with ARD Maintenance}
\usage{
sim_wiener_maintenance_paths(
  n_units,
  t_max,
  n_steps,
  drift,
  sigma2,
  rho,
  n_maint
)
}
\arguments{
\item{n_units}{Integer. Number of independent units/systems to simulate.}

\item{t_max}{Numeric. Maximum observation time.}

\item{n_steps}{Integer. Number of measurement intervals.}

\item{drift}{Numeric. Drift parameter (\eqn{\nu} or \eqn{\mu}).}

\item{sigma2}{Numeric. Diffusion variance parameter (\eqn{\sigma^2}).}

\item{rho}{Numeric vector or scalar. Maintenance efficiency parameters for each intervention (\eqn{0 \le \rho_k \le 1}).}

\item{n_maint}{Integer. Number of maintenance interventions.}
}
\value{
A \code{data.frame} containing simulated degradation paths for all units.
}
\description{
Simulates multiple independent Wiener process paths subjected to imperfect maintenance
actions under the Arithmetic Reduction of Degradation (ARD) model.
}
\examples{
data_paths <- sim_wiener_maintenance_paths(
  n_units = 3, t_max = 20, n_steps = 20, drift = 2, sigma2 = 2,
  rho = c(0.8, 0.5, 0.8), n_maint = 3
)
head(data_paths)

}
